+++ __tests__/e2e/createShortUrl.spec.ts
@@ -4,6 +4,8 @@ import { ExpressApp } from '@infra/http/express';
 import { RateLimit } from '@infra/middlewares/rate-limit/rateLimit';
 import { PinoLogger } from '@infra/listeners/loggers/pinoLogger';
 import { HttpResponse } from '@helpers/httpResponse';
+import { CreateShortenedUrl } from '@modules/shortenedUrls/controllers/createShortenedUrl';
+import { ioRedis } from '@infra/db/redis/ioRedisHelper';
 
 const app = new ExpressApp();
 app.setupRoutes();
@@ -18,6 +20,18 @@ describe('Create short url', () => {
 		jest.spyOn(PinoLogger.prototype, 'info').mockImplementation();
 		jest.spyOn(PinoLogger.prototype, 'warn').mockImplementation();
 		jest.spyOn(PinoLogger.prototype, 'error').mockImplementation();
+		jest.spyOn(
+			CreateShortenedUrl.prototype as any,
+			'sendToShortenedUrlCreationQueue',
+		).mockImplementation();
+		jest.spyOn(
+			CreateShortenedUrl.prototype as any,
+			'createLongTermCache',
+		).mockImplementation();
+	});
+
+	afterAll(async () => {
+		await ioRedis.quit();
 	});
 
 	test('Should return status code 400 when url is not passed in request body', async () => {