+++ src/infra/cache/ICache.ts
@@ -1,5 +1,10 @@
 export interface ICache {
 	set(key: string, value: string): Promise<void>;
+	setWithExpiration(
+		key: string,
+		value: string,
+		expireTime: number,
+	): Promise<void>;
 	get(key: string): Promise<string | null>;
 	del(key: string): Promise<void>;
 }