+++ src/infra/db/redis/ioRedisHelper.ts
@@ -1,4 +1,4 @@
 import { variables } from '@helpers/envs';
-import IORedis from 'ioredis';
+import Redis from 'ioredis';
 
-export const ioRedis = new IORedis(variables.redisUrl);
+export const ioRedis = new Redis(variables.redisUrl);